backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pymupdf==1.24.11
chromadb==0.5.18
google-generativeai==0.8.3
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic[email]==2.9.2